ml-service/requirements.txt
flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.6.0
yfinance==0.1.63
python-dotenv==0.19.0
requests==2.26.0
pandas-ta==0.3.14b0
prophet==1.0.1
